Restructure tests to make them importable from other packages
