+++ setup.py
@@ -66,7 +66,7 @@ setup(
     ],
     packages=find_packages(exclude=['tests*', 'dev*']),
 
-    test_suite='dev.tests.make_suite',
+    test_suite='tests.make_suite',
 
     cmdclass={
         'clean': CleanCommand,
